refactor: move some code around
